api-auth/requirements.txt
flask
flask_restful
flask_sqlalchemy
psycopg2
pyjwt
webargs
